Trait narwhal_dag::node_dag::Affiliated

pub trait Affiliated: Hash<{ crypto::DIGEST_LENGTH }> {
    fn parents(
        &self
    ) -> Vec<<Self as Hash<{ crypto::DIGEST_LENGTH }>>::TypedDigest> ;

    fn compressible(&self) -> bool { ... }
}

A trait marking the minimum information we need to sort out the value for a node:

parents: hash pointers to its parents, and

compressible: a value-derived boolean indicating if that value is, initially, compressible

The crypto:Hash trait bound offers the digest-ibility.

Required Methods

fn parents(&self) -> Vec<<Self as Hash<{ crypto::DIGEST_LENGTH }>>::TypedDigest>

Hash pointers to the parents of the current value

Provided Methods

fn compressible(&self) -> bool

Whether the current value should be marked as compressible when first inserted in a Node. Defaults to a blanket false for all values.